move refactored function to @tinacms/schema-tools
